Reorganize example JS
Clustering everything into chunks that need to be handled at about the same
time. I got rid of all the anonymous functions so everything looks the same and
to avoid all the extra syntax that's prone to typos.

Also got rid of the connect event callback since it's non-critical. Will add it
as the first extra work item.