Update README to match new example code
I gave in and just included the publish/subscribe keys in the code. It's easier
than making everyone copy it or fetch it separately, and it's not like they're
really secrets to anyone visiting the page.

I think this order of operations makes pretty good sense, and lets us test the
interesting functions in chunks as we go.